refactor: 중복 코드 제거를 위한 합성 사용
