+++ src/main/java/model/piece/Chariot.java
@@ -2,7 +2,6 @@ package model.piece;
 
 import static model.Movement.*;
 
-import java.util.ArrayList;
 import java.util.List;
 import model.Movement;
 import model.Team;
@@ -11,9 +10,11 @@ import model.position.Position;
 public class Chariot extends Piece {
 
     private final List<Movement> movements = List.of(UP, DOWN, LEFT, RIGHT);
+    private final MovableNavigator movableNavigator;
 
     public Chariot(Team team) {
         super(team);
+        movableNavigator = new UnLimitedCrossNavigator();
     }
 
     @Override
@@ -31,19 +32,6 @@ public class Chariot extends Piece {
 
     @Override
     public List<Position> calculateAllDirection(Position departure, Position arrival) {
-        return movements.stream()
-            .map(movement -> findDirectionByMovement(departure, arrival, movement))
-            .filter(moveDirections -> moveDirections.contains(arrival))
-            .findFirst()
-            .orElseThrow(() -> new IllegalArgumentException("해당 위치로 이동할 수 없습니다."));
-    }
-
-    private List<Position> findDirectionByMovement(Position departure, Position arrival, Movement movement) {
-        List<Position> movedPosition = new ArrayList<>();
-        while (departure.canMove(movement) && !movedPosition.contains(arrival)) {
-            departure = departure.move(movement);
-            movedPosition.add(departure);
-        }
-        return movedPosition;
+        return movableNavigator.find(departure, arrival, movements);
     }
 }
