+++ src/main/java/model/piece/Piece.java
@@ -4,7 +4,7 @@ import java.util.List;
 import model.Team;
 import model.position.Position;
 
-public abstract class Piece{
+public abstract class Piece {
 
     private final Team team;
